Lib: Added dbidAsParam, fixed UserRoles
API_GetAppDTMInfo needed to be executed against /db/main with the dbid
passed in as a query string parameter. To achieve this, a new flag in the
settings.flags object has been created: dbidAsParam (defaults to false).

When dbidAsParam is true, any passed in dbid option will be sent via the
query string or xml packet instead of the URL path.

This is done automatically for API_GetAppDTMInfo.

API_UserRoles was not being properly normalized. This commit normalizes
the users object and the user.roles object.